+++ cmn/conf.c
@@ -64,7 +64,7 @@ bool load_peer_conf(const char *pConfFile, peer_conf_t *pPeerConf)
     memset(pPeerConf, 0, sizeof(peer_conf_t));
 
     if (ini_parse(pConfFile, handler_peer_conf, pPeerConf) != 0) {
-        SYSLOG_ERR("fail peer parse[%s]", pConfFile);
+        //DBG_PRINTF("fail peer parse[%s]", pConfFile);
         return false;
     }
 
@@ -96,7 +96,7 @@ bool load_funding_conf(const char *pConfFile, funding_conf_t *pFundConf)
     memset(pFundConf, 0, sizeof(funding_conf_t));
 
     if (ini_parse(pConfFile, handler_fund_conf, pFundConf) != 0) {
-        SYSLOG_ERR("fail fund parse[%s]", pConfFile);
+        //DBG_PRINTF("fail fund parse[%s]", pConfFile);
         return false;
     }
 
@@ -129,12 +129,12 @@ void print_funding_conf(const funding_conf_t *pFundConf)
 bool load_btcrpc_conf(const char *pConfFile, rpc_conf_t *pRpcConf)
 {
     if (ini_parse(pConfFile, handler_btcrpc_conf, pRpcConf) != 0) {
-        SYSLOG_ERR("fail bitcoin.conf parse[%s]", pConfFile);
+        DBG_PRINTF("fail bitcoin.conf parse[%s]", pConfFile);
         return false;
     }
 
     if ((strlen(pRpcConf->rpcuser) == 0) || (strlen(pRpcConf->rpcpasswd) == 0)) {
-        SYSLOG_ERR("fail: no rpcuser or rpcpassword[%s]", pConfFile);
+        DBG_PRINTF("fail: no rpcuser or rpcpassword[%s]", pConfFile);
         return false;
     }
 
@@ -165,7 +165,7 @@ bool load_payment_conf(const char *pConfFile, payment_conf_t *pPayConf)
     memset(pPayConf, 0, sizeof(payment_conf_t));
 
     if (ini_parse(pConfFile, handler_pay_conf, pPayConf) != 0) {
-        SYSLOG_ERR("fail pay parse[%s]", pConfFile);
+        DBG_PRINTF("fail pay parse[%s]", pConfFile);
         return false;
     }
 
@@ -203,7 +203,7 @@ bool load_anno_conf(const char *pConfFile, anno_conf_t *pAnnoConf)
     memset(pAnnoConf, 0, sizeof(anno_conf_t));
 
     if (ini_parse(pConfFile, handler_anno_conf, pAnnoConf) != 0) {
-        SYSLOG_ERR("fail anno parse[%s]", pConfFile);
+        //DBG_PRINTF("fail anno parse[%s]", pConfFile);
         return false;
     }
 
@@ -216,7 +216,7 @@ bool load_establish_conf(const char *pConfFile, establish_conf_t *pEstConf)
     memset(pEstConf, 0, sizeof(establish_conf_t));
 
     if (ini_parse(pConfFile, handler_establish_conf, pEstConf) != 0) {
-        SYSLOG_ERR("fail establish parse[%s]", pConfFile);
+        //DBG_PRINTF("fail establish parse[%s]", pConfFile);
         return false;
     }