+++ ucoind/ucoind.c
@@ -186,7 +186,7 @@ int main(int argc, char *argv[])
 
     //peer config出力
     char fname[256];
-    sprintf(fname, "ptarm_%s.conf", p_alias);
+    sprintf(fname, FNAME_FMT_NODECONF, p_alias);
     FILE *fp = fopen(fname, "w");
     if (fp) {
         if (p_addr->type == LN_NODEDESC_IPV4) {